refactor(queues): extract shared negotiation timeout core
Collapse the byte-identical buildOutcome + the AI-fallback/finalize/re-arm
block duplicated between NegotiationTimeoutQueue and NegotiationClaimTimeoutQueue
into timeout.shared.ts (buildNegotiationOutcome + runTimeoutFallback). Each
worker now keeps only its distinct task-acquisition (state guard vs atomic
claimed->working transition), labels, seed reasoning, max-turns source, and
re-arm strategy. Characterization tests (13) stay green; jscpd duplication
between the two files dropped from ~140 to 37 lines (BullMQ scaffolding).
