+++ backend/src/queues/negotiations/timeout.queue.ts
@@ -2,8 +2,10 @@ import { Job } from 'bullmq';
 import { log } from '../../lib/log';
 import { QueueFactory } from '../../lib/bullmq/bullmq';
 import { conversationDatabaseAdapter } from '../../adapters/database.adapter';
-import { IndexNegotiator, AMBIENT_PARK_WINDOW_MS } from '@indexnetwork/protocol';
-import type { NegotiationTurn, NegotiationOutcome, UserNegotiationContext, SeedAssessment, NegotiationGraphDatabase } from '@indexnetwork/protocol';
+import { AMBIENT_PARK_WINDOW_MS } from '@indexnetwork/protocol';
+import type { NegotiationGraphDatabase } from '@indexnetwork/protocol';
+
+import { runTimeoutFallback, type NegotiationTaskMeta } from './timeout.shared';
 
 /** BullMQ queue name for negotiation timeout jobs. */
 export const QUEUE_NAME = 'negotiation-timeout';
@@ -177,143 +179,33 @@ export class NegotiationTimeoutQueue {
       return;
     }
 
-    const meta = task.metadata as { sourceUserId?: string; candidateUserId?: string; type?: string } | null;
+    const meta = task.metadata as NegotiationTaskMeta | null;
     if (meta?.type !== 'negotiation') {
       this.logger.warn('[NegotiationTimeoutJob] Task is not a negotiation, skipping', { negotiationId });
       return;
     }
 
-    // Determine whose turn it is
-    const currentSpeaker = currentTurnCount % 2 === 0 ? 'source' : 'candidate';
-    const isSource = currentSpeaker === 'source';
-    const activeUserId = isSource ? meta.sourceUserId! : meta.candidateUserId!;
-    const otherUserId = isSource ? meta.candidateUserId! : meta.sourceUserId!;
-
-    this.logger.info('[NegotiationTimeoutJob] External agent timed out, running AI fallback', {
+    await runTimeoutFallback({
+      database,
+      logger: this.logger,
+      labels: {
+        job: '[NegotiationTimeoutJob]',
+        fallback: 'External agent timed out, running AI fallback',
+        finalized: 'Negotiation finalized after timeout',
+        statusUpdateFailed: 'Failed to update opportunity status on timeout finalization',
+      },
       negotiationId,
-      activeUserId,
-      turnNumber,
-    });
-
-    // Parse history
-    const history: NegotiationTurn[] = messages.map((m: { parts: unknown[] }) => {
-      const dp = (m.parts as Array<{ kind?: string; data?: unknown }>)?.find(p => p.kind === 'data');
-      return dp?.data as NegotiationTurn;
-    }).filter(Boolean);
-
-    // Run AI agent for the timed-out turn
-    const agent = new IndexNegotiator();
-    const ownUserCtx: UserNegotiationContext = { id: activeUserId, intents: [], profile: {} };
-    const otherUserCtx: UserNegotiationContext = { id: otherUserId, intents: [], profile: {} };
-    const seedAssessment: SeedAssessment = { reasoning: 'Timeout fallback', valencyRole: 'peer' };
-
-    const aiTurn = await agent.invoke({
-      ownUser: ownUserCtx,
-      otherUser: otherUserCtx,
-      indexContext: { networkId: '', prompt: '' },
-      seedAssessment,
-      history,
-      isDiscoverer: isSource,
-    });
-
-    // Persist the AI turn
-    await database.createMessage({
-      conversationId: task.conversationId,
-      senderId: `agent:${activeUserId}`,
-      role: 'agent',
-      parts: [{ kind: 'data' as const, data: aiTurn }],
       taskId: task.id,
+      conversationId: task.conversationId,
+      meta,
+      messages,
+      currentTurnCount,
+      seedReasoning: 'Timeout fallback',
+      maxTurns: 6,
+      rearm: async (newTurnCount) => {
+        await this.enqueueTimeout(negotiationId, newTurnCount, AMBIENT_PARK_WINDOW_MS);
+      },
     });
-
-    const newTurnCount = currentTurnCount + 1;
-    const maxTurns = 6;
-
-    // Evaluate: accept/reject → finalize; counter at max → finalize; counter under max → continue
-    if (aiTurn.action === 'accept' || aiTurn.action === 'reject' || newTurnCount >= maxTurns) {
-      const fullHistory = [...history, aiTurn];
-      const nextSpeaker = currentSpeaker === 'source' ? 'candidate' : 'source';
-      const outcome = this.buildOutcome(fullHistory, newTurnCount, aiTurn.action, meta.sourceUserId!, meta.candidateUserId!, nextSpeaker);
-
-      await database.updateTaskState(task.id, 'completed');
-      await database.createArtifact({
-        taskId: task.id,
-        name: 'negotiation-outcome',
-        parts: [{ kind: 'data', data: outcome }],
-        metadata: { hasOpportunity: outcome.hasOpportunity, turnCount: newTurnCount },
-      });
-
-      const outcomeStr = aiTurn.action === 'accept' ? 'accepted'
-        : aiTurn.action === 'reject' ? 'rejected'
-        : 'turn_cap';
-
-      const opportunityId = (meta as { opportunityId?: string }).opportunityId;
-      if (opportunityId) {
-        const nextStatus = aiTurn.action === 'accept' ? 'pending'
-          : aiTurn.action === 'reject' ? 'rejected'
-          : 'stalled';
-        await database.updateOpportunityStatus(opportunityId, nextStatus).catch((err) => {
-          this.logger.error('[NegotiationTimeoutJob] Failed to update opportunity status on timeout finalization', {
-            opportunityId,
-            nextStatus,
-            error: err,
-          });
-        });
-      }
-
-      this.logger.info('[NegotiationTimeoutJob] Negotiation finalized after timeout', {
-        negotiationId,
-        outcome: outcomeStr,
-        turnCount: newTurnCount,
-      });
-      return;
-    }
-
-    // AI countered and under max turns — the other party now needs to respond.
-    // Set to waiting_for_agent and arm a new timeout so the negotiation doesn't stall.
-    await database.updateTaskState(task.id, 'waiting_for_agent');
-
-    await this.enqueueTimeout(negotiationId, newTurnCount, AMBIENT_PARK_WINDOW_MS);
-
-    this.logger.info('[NegotiationTimeoutJob] AI agent countered, armed timeout for next speaker', {
-      negotiationId,
-      action: aiTurn.action,
-      turnCount: newTurnCount,
-    });
-  }
-
-  /** Build a NegotiationOutcome (mirrors graph finalizeNode logic). */
-  private buildOutcome(
-    history: NegotiationTurn[],
-    turnCount: number,
-    lastAction: string,
-    sourceUserId: string,
-    candidateUserId: string,
-    currentSpeaker: string,
-  ): NegotiationOutcome {
-    const hasOpportunity = lastAction === 'accept';
-    const atCap = lastAction === 'counter';
-
-    let agreedRoles: NegotiationOutcome['agreedRoles'] = [];
-    if (hasOpportunity && history.length >= 2) {
-      const acceptTurn = history[history.length - 1];
-      const precedingTurn = history[history.length - 2];
-      const accepterIsSource = currentSpeaker === 'candidate';
-      const [sourceRole, candidateRole] = accepterIsSource
-        ? [acceptTurn.assessment.suggestedRoles.ownUser, precedingTurn.assessment.suggestedRoles.ownUser]
-        : [precedingTurn.assessment.suggestedRoles.ownUser, acceptTurn.assessment.suggestedRoles.ownUser];
-      agreedRoles = [
-        { userId: sourceUserId, role: sourceRole },
-        { userId: candidateUserId, role: candidateRole },
-      ];
-    }
-
-    return {
-      hasOpportunity,
-      agreedRoles,
-      reasoning: history[history.length - 1]?.assessment.reasoning ?? '',
-      turnCount,
-      ...(atCap && { reason: 'turn_cap' }),
-    };
   }
 }
